requirements.txt
absl-py==2.1.0
annotated-types 
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.2.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
backcall==0.2.0
bcrypt==4.1.3
beautifulsoup4==4.12.3
bleach==6.1.0
cachetools 
cffi==1.16.0
click==8.1.7
colorama==0.4.6
comm==0.1.4
contourpy==1.2.1
cryptography==42.0.7
cycler==0.12.1
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
ecdsa==0.19.0
entrypoints==0.4
executing==1.2.0
fastapi 
fastjsonschema==2.19.1
flatbuffers==24.3.25
fonttools==4.51.0
fqdn==1.5.1
greenlet 
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
idna 
ipykernel==6.25.0
ipython==8.14.0
isoduration==20.11.0
jax==0.4.28
jaxlib==0.4.28
jedi==0.19.0
Jinja2==3.1.3
json5==0.9.14
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==7.4.9
jupyter_core==5.3.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.1.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.3
kiwisolver==1.4.5
matplotlib==3.8.4
matplotlib-inline==0.1.6
mediapipe==0.10.14
mistune==3.0.2
ml-dtypes==0.4.0
nbclient==0.9.0
nbconvert==7.16.0
nbformat==5.9.2
nest-asyncio==1.5.7
notebook==7.1.0
notebook_shim==0.2.4
numpy==1.26.4
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opt-einsum==3.3.0
overrides==7.7.0
packaging==23.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.3
passlib==1.7.4
pickleshare==0.7.5
pillow==10.3.0
platformdirs==3.10.0
prometheus_client==0.20.0
prompt-toolkit==3.0.39
protobuf==4.25.3
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.6.0
pycparser==2.21
pydantic 
pydantic_core 
Pygments==2.15.1
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyzmq==24.0.1
referencing==0.33.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.0
rsa==4.9
scipy==1.13.0
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
sounddevice==0.4.6
soupsieve==2.5
SQLAlchemy 
stack-data==0.6.2
starlette 
terminado==0.18.0
tinycss2==1.2.1
tornado==6.3.2
traitlets==5.9.0
types-python-dateutil==2.8.19.20240106
typing_extensions 
tzdata==2024.1
uri-template==1.3.0
uvicorn==0.29.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
